requirements.txt
streamlit
pandas
python-dotenv
httpx
langchain
langchain_openai
regex